Dev/bootcamp/requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
django-crispy-forms==1.6.0
django-errortemplates==1.0.1
whitenoise==5.2.0
django-paypal==1.1
pillow==8.4.0
djangorestframework==3.13.1
